backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
marshmallow==3.20.1
reportlab==4.0.7
python-dotenv==1.0.0
pytest==7.4.3
gunicorn==21.2.0
email-validator==2.1.0
